requirements.txt
streamlit
openai
langchain
tiktoken
python-dotenv
google-cloud-aiplatform
google-auth
google-generativeai